Cleanup components * Renaming spec files * Mergin non standalone components
