+++ src/components/content-box/ContentBox.spec.jsx
