+++ src/components/icon-as-button/IconAsButton.spec.jsx
