+++ src/components/subject-icons/SubjectIcon.spec.jsx
